+++ crates/client/metrics/src/block_metrics/block_metrics.rs
@@ -1,4 +1,6 @@
-use dc_metrics::{Gauge, MetricsRegistry, PrometheusError, F64};
+// use dc_metrics::{Gauge, MetricsRegistry, PrometheusError, F64};
+
+use crate::{MetricsRegistry, PrometheusError, Gauge, F64};
 
 #[derive(Clone, Debug)]
 pub struct BlockMetrics {
