+++ crates/client/metrics/src/block_metrics/mod.rs
@@ -1,0 +1,1 @@
+pub mod block_metrics;
