+++ crates/client/sync/Cargo.toml
@@ -20,6 +20,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 # Deoxys
 dc-db = { workspace = true }
+dc-eth = {workspace = true}
 dc-metrics = { workspace = true }
 dc-telemetry = { workspace = true }
 dp-block = { workspace = true }
@@ -42,6 +43,7 @@ starknet_api = { workspace = true }
 anyhow = "1.0.75"
 bitvec = { workspace = true }
 ethers = { workspace = true }
+alloy = { workspace = true }
 futures = { workspace = true, default-features = true }
 hex = { workspace = true }
 log = { workspace = true }
