+++ crates/client/sync/src/lib.rs
@@ -2,9 +2,7 @@
 
 pub mod commitments;
 pub mod fetch;
-pub mod l1;
 pub mod l2;
-pub mod metrics;
 pub mod reorgs;
 pub mod utils;
 
@@ -18,23 +16,22 @@ use starknet_types_core::felt::Felt;
 
 pub mod starknet_sync_worker {
     use std::{sync::Arc, time::Duration};
-
+    use alloy::primitives::Address;
     use anyhow::Context;
     use dc_db::{db_metrics::DbMetrics, DeoxysBackend};
     use dc_telemetry::TelemetryHandle;
     use reqwest::Url;
     use starknet_providers::SequencerGatewayProvider;
-
+    use dc_metrics::block_metrics::block_metrics::BlockMetrics;
     use self::fetch::fetchers::FetchConfig;
     use super::*;
-    use crate::metrics::block_metrics::BlockMetrics;
 
     #[allow(clippy::too_many_arguments)]
     pub async fn sync(
         backend: &Arc<DeoxysBackend>,
         fetch_config: FetchConfig,
         l1_url: Option<Url>,
-        l1_core_address: ethers::abi::Address,
+        l1_core_address: Address,
         starting_block: Option<u64>,
         backup_every_n_blocks: Option<u64>,
         block_metrics: BlockMetrics,
@@ -70,7 +67,7 @@ pub mod starknet_sync_worker {
         let l1_block_metric = block_metrics.clone();
         let l1_fut = async {
             if let Some(l1_url) = l1_url {
-                l1::sync(backend, l1_url.clone(), l1_block_metric, l1_core_address, chain_id).await
+                dc_eth::state_update::sync(backend, l1_url.clone(), l1_block_metric, l1_core_address, chain_id).await
             } else {
                 Ok(())
             }
